feat:CRUD de usuario e tatuador
