+++ API-Ink-4-You/ink-4-you/src/main/java/br/com/bandtec/ink4yousembanco/model/Usuario.java
@@ -1,0 +1,34 @@
+package br.com.bandtec.ink4yousembanco.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import net.bytebuddy.dynamic.loading.InjectionClassLoader;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import java.util.Date;
+
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Entity
+public class Usuario {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id_usuario;
+
+    private String nome;
+    private Date data_nascimento;
+    private String cpf;
+    private String cep;
+    private String telefone;
+    private String email;
+    private String senha;
+    private String foto_perfil;
+}
+
