--- API-sem-conexao-banco/ink-4-you-sem-banco/src/main/java/br/com/bandtec/ink4yousembanco/Ink4YouSemBancoApplication.java
@@ -1,34 +1,0 @@
-package br.com.bandtec.ink4yousembanco;
-
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-import java.util.stream.LongStream;
-
-@SpringBootApplication
-public class Ink4YouSemBancoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Ink4YouSemBancoApplication.class, args);
-	}
-
-//	@Bean
-//	CommandLineRunner init(ClienteRepository repository) {
-//		return args -> {
-//			repository.deleteAll();
-//			LongStream.range(1, 11)
-//					.mapToObj(i -> {
-//						Contact c = new Contact();
-//						c.setName("Contact " + i);
-//						c.setEmail("contact" + i + "@email.com");
-//						c.setPhone("(111) 111-1111");
-//						return c;
-//					})
-//					.map(v -> repository.save(v))
-//					.forEach(System.out::println);
-//		};
-//	}
-
-}
