--- API-sem-conexao-banco/ink-4-you-sem-banco/src/main/java/br/com/bandtec/ink4yousembanco/model/Cliente.java
@@ -1,22 +1,0 @@
-package br.com.bandtec.ink4yousembanco.model;
-
-import java.util.Date;
-
-public class Cliente extends Usuario {
-
-    private String cpf;
-
-    public Cliente(String email, String nome, String senha,
-                   long celular, String cep, Date dataNascimento, String cpf, String fotoPerfil) {
-        super(email, nome, senha, celular, cep, dataNascimento, fotoPerfil);
-        this.cpf = cpf;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-}
