--- API-sem-conexao-banco/ink-4-you-sem-banco/src/main/java/br/com/bandtec/ink4yousembanco/repository/ClienteRepository.java
@@ -1,8 +1,0 @@
-package br.com.bandtec.ink4yousembanco.repository;
-
-import br.com.bandtec.ink4yousembanco.model.Cliente;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {}
