--- API-sem-conexao-banco/ink-4-you-sem-banco/src/main/java/br/com/bandtec/ink4yousembanco/repository/TatuadorRepository.java
@@ -1,9 +1,0 @@
-package br.com.bandtec.ink4yousembanco.repository;
-
-import br.com.bandtec.ink4yousembanco.model.Cliente;
-import br.com.bandtec.ink4yousembanco.model.Tatuador;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TatuadorRepository extends JpaRepository<Tatuador, Long> {}
