--- API-sem-conexao-banco/ink-4-you-sem-banco/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-
