--- POC-API-BANCO/poc-banco/src/main/java/br/com/bandtec/pocbanco/PocBancoApplication.java
@@ -1,35 +1,0 @@
-package br.com.bandtec.pocbanco;
-
-import br.com.bandtec.pocbanco.model.Contact;
-import br.com.bandtec.pocbanco.repository.ContactRepository;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-
-import java.util.stream.LongStream;
-
-@SpringBootApplication
-public class PocBancoApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PocBancoApplication.class, args);
-	}
-
-	@Bean
-	CommandLineRunner init(ContactRepository repository) {
-		return args -> {
-			repository.deleteAll();
-			LongStream.range(1, 11)
-					.mapToObj(i -> {
-						Contact c = new Contact();
-						c.setName("Contact " + i);
-						c.setEmail("contact" + i + "@email.com");
-						c.setPhone("(111) 111-1111");
-						return c;
-					})
-					.map(v -> repository.save(v))
-					.forEach(System.out::println);
-		};
-	}
-}
