--- POC-API-BANCO/poc-banco/src/main/java/br/com/bandtec/pocbanco/model/Contact.java
@@ -1,24 +1,0 @@
-package br.com.bandtec.pocbanco.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-@Entity
-public class Contact {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String name;
-    private String email;
-    private String phone;
-}
