--- POC-API-BANCO/poc-banco/src/main/java/br/com/bandtec/pocbanco/repository/ContactRepository.java
@@ -1,8 +1,0 @@
-package br.com.bandtec.pocbanco.repository;
-
-import br.com.bandtec.pocbanco.model.Contact;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ContactRepository extends JpaRepository<Contact, Long> {}
