--- POC-API-BANCO/poc-banco/src/main/resources/application.properties
@@ -1,11 +1,0 @@
-
-## Spring DATASOURCE (DataSourceAutoConfiguration & DataSourceProperties)
-spring.datasource.url=jdbc:sqlserver://magic-site.database.windows.net:1433;database=Poc-Banco;user=Chris@magic-site;password={#Gf48635896890};encrypt=true;trustServerCertificate=false;hostNameInCertificate=*.database.windows.net;loginTimeout=30;
-spring.datasource.username=Chris
-spring.datasource.password=#Gf48635896890
-
-# Dialeto SQL melhorar o SQL gerado pelo Hibernate
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5InnoDBDialect
-
-# Hibernate ddl auto (create, create-drop, validate, update)
-spring.jpa.hibernate.ddl-auto=update
