--- POC-API-BANCO/poc-banco/src/test/java/br/com/bandtec/pocbanco/PocBancoApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.bandtec.pocbanco;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PocBancoApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
