--- ink-4-you-2/src/main/java/br/com/bandtec/ink4you2/Autenticavel.java
@@ -1,10 +1,0 @@
-package br.com.bandtec.ink4you2;
-
-public interface Autenticavel {
-
-    public void cadastrarUsuario();
-
-    public void logar(String email, String senha);
-
-    public void logoff(Integer userID);
-}
