--- ink-4-you-2/src/main/resources/application.properties
@@ -1,7 +1,0 @@
-logging.level.org.springframework.jdbc.core=DEBUG
-
-spring.datasource.url=jdbc:sqlserver://ink4you.database.windows.net:1433;database=bd-ink4you;user=ink4you@ink4you;password={#Gfgrupo8};encrypt=true;trustServerCertificate=false;hostNameInCertificate=*.database.windows.net;loginTimeout=30;
-spring.datasource.username=ink4you
-spring.datasource.password=#Gfgrupo8
-
-spring.datasource.initialization-mode=always
