--- ink-4-you-2/src/test/java/br/com/bandtec/ink4you2/Ink4You2ApplicationTests.java
@@ -1,13 +1,0 @@
-package br.com.bandtec.ink4you2;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Ink4You2ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
